feat(develop): Add `npm tdd` and `grunt tdd` tasks
Added a plugin to continue grunt process on karma on failure, used by
a new grunt task that just watches, compiles typescript, & runs karma,
The goal is helping contributers write unit tests, even a
test-driven workflow.

Also an `npm` script, `npm tdd` and a note about it in the docs.